requirements.txt
pyzbar
Pillow>=5.0
qrcode>=6.0